added the watch command (previously known as watcher --start)
current implementation is same as I implemented in python this can be
improved further using goroutines when implementing AFK feature.
This implementation is without AFK feature as of now. For AFK feature
we can think of two go routine one for writing into the db and other for
Checking and updating the user's AFK status in channel